move compliance check to fx
